// manager_frontend/src/client/models/ManagerStorefrontResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ManagerStorefrontResponse = {
    slug: string;
    display_name: string;
    city?: (string | null);
    default_locale: string;
    currency: string;
    is_default: boolean;
    is_current: boolean;
};